refactor: рефакторинг страницы Rights по замечаниям
