+++ src/components/CardsSectionWithLines/CardsSectionWithLines.js
@@ -4,7 +4,7 @@ import { Loader, Paginate } from '../utils';
 import renderThoseDamnedLines from '../../utils/render-lines';
 import './CardsSectionWithLines.scss';
 
-const maxTabletWidth = '900px';
+const maxTabletWidth = 900;
 
 function CardsSectionWithLines({
   pageCount,
@@ -19,7 +19,7 @@ function CardsSectionWithLines({
   const [isTablet, setIsTablet] = useState(false);
 
   useEffect(() => {
-    const tablet = window.matchMedia(`(max-width: ${maxTabletWidth})`);
+    const tablet = window.matchMedia(`(max-width: ${maxTabletWidth}px)`);
 
     const listener = () => {
       if (tablet.matches) setIsTablet(true);
