+++ src/pages/Places/Places.js
@@ -53,7 +53,7 @@ const {
   mentorTag,
 } = placesPageTexts;
 
-const smallQueryWidth = '1024px';
+const smallQueryWidth = 1024;
 
 function Places() {
   const { state } = useLocation();
@@ -102,7 +102,7 @@ function Places() {
 
   // Резайз пагинации при первой загрузке
   useEffect(() => {
-    const smallQuery = window.matchMedia(`(max-width: ${smallQueryWidth})`);
+    const smallQuery = window.matchMedia(`(max-width: ${smallQueryWidth}px)`);
 
     const listener = () => {
       if (smallQuery.matches) {
